[MMCA-4369] Move EmailVerified and EmainUnverified response to EmailResponses
